+++ gradle.properties
@@ -37,3 +37,4 @@ android.r8.optimizedResourceShrinking=true
 android.generateSyncIssueWhenLibraryConstraintsAreEnabled=false
 android.builtInKotlin=true
 android.newDsl=true
+org.gradle.tooling.parallel=true
