+++ readium/navigator/src/main/java/org/readium/r2/navigator/epub/EpubNavigatorFragment.kt
@@ -80,6 +80,7 @@ import org.readium.r2.navigator.pager.R2ViewPager
 import org.readium.r2.navigator.preferences.Configurable
 import org.readium.r2.navigator.preferences.FontFamily
 import org.readium.r2.navigator.preferences.ReadingProgression
+import org.readium.r2.navigator.toTargetElement
 import org.readium.r2.navigator.util.createFragmentFactory
 import org.readium.r2.shared.DelicateReadiumApi
 import org.readium.r2.shared.ExperimentalReadiumApi
@@ -90,7 +91,6 @@ import org.readium.r2.shared.publication.Layout
 import org.readium.r2.shared.publication.Link
 import org.readium.r2.shared.publication.Locator
 import org.readium.r2.shared.publication.Publication
-import org.readium.r2.shared.publication.services.content.Content
 import org.readium.r2.shared.publication.services.positionsByReadingOrder
 import org.readium.r2.shared.util.AbsoluteUrl
 import org.readium.r2.shared.util.Url
@@ -738,73 +738,6 @@ public class EpubNavigatorFragment internal constructor(
             RectF(left, top + topOffset, right, bottom)
         } ?: this
 
-    /**
-     * Builds the public [TargetElement] from the raw metadata produced by gestures.js.
-     */
-    private fun buildTargetElement(data: TargetElementData): TargetElement? {
-        // The resource owning the web view that emitted the tap.
-        val resourceLink = currentReflowablePageFragment?.link ?: return null
-
-        // Locator pointing to the element inside the resource that contains it.
-        val locator = Locator(
-            href = resourceLink.url(),
-            mediaType = resourceLink.mediaType ?: MediaType.XHTML,
-            locations = Locator.Locations(
-                otherLocations = data.cssSelector
-                    ?.let { mapOf("cssSelector" to it) }
-                    ?: emptyMap()
-            )
-        )
-
-        val content = buildContentElement(data, locator) ?: return null
-
-        return TargetElement(
-            frame = data.frame.adjustedToViewport(),
-            content = content
-        )
-    }
-
-    /**
-     * Maps the target element metadata to a [Content.Element], following the same discrimination
-     * rule as the Swift toolkit: when a resolvable source is found, the element is an
-     * [Content.ImageElement] (covers `<img>` and `<svg href=...>`); otherwise an inline `<svg>`
-     * becomes a [Content.SvgElement].
-     */
-    private fun buildContentElement(data: TargetElementData, locator: Locator): Content.Element? {
-        val attributes = buildList {
-            data.accessibilityLabel?.takeIf { it.isNotBlank() }?.let { label ->
-                add(Content.Attribute(Content.AttributeKey.ACCESSIBILITY_LABEL, label))
-            }
-        }
-
-        // Look up the source in the publication manifest so the client gets full metadata (media
-        // type, etc.). For resources not in the manifest (e.g. external images), we synthesise a
-        // plain Link.
-        val embeddedLink = (data.src?.let { Url(it) } as? AbsoluteUrl)
-            ?.let { url -> viewModel.internalLinkFromUrl(url) ?: Link(href = url) }
-
-        if (embeddedLink != null) {
-            return Content.ImageElement(
-                locator = locator,
-                embeddedLink = embeddedLink,
-                caption = data.caption,
-                attributes = attributes
-            )
-        }
-
-        // Inline SVG fallback.
-        if (data.tag == "svg" && data.html != null) {
-            return Content.SvgElement(
-                locator = locator,
-                svg = data.html,
-                caption = data.caption,
-                attributes = attributes
-            )
-        }
-
-        return null
-    }
-
     // DecorableNavigator
 
     override fun <T : Decoration.Style> supportsDecorationStyle(style: KClass<T>): Boolean =
@@ -864,6 +797,15 @@ public class EpubNavigatorFragment internal constructor(
             inputListener.onTap(
                 TapEvent(point, targetElement?.let { buildTargetElement(it) })
             )
+
+        private fun buildTargetElement(data: TargetElementData): TargetElement? =
+            currentReflowablePageFragment?.link?.let { resourceLink ->
+                data.toTargetElement(
+                    resourceLink = resourceLink,
+                    adjustFrame = { it.adjustedToViewport() },
+                    internalLinkForUrl = viewModel::internalLinkFromUrl
+                )
+            }
 
         override fun onDragStart(event: R2BasicWebView.DragEvent): Boolean =
             onDrag(DragEvent.Type.Start, event)
